refactor: move compairing section from shared to pricing
