+++ src/components/pages/pricing/compairing/index.ts
