+++ src/components/pages/pricing/compairing/table/index.ts
